# pkg/api/docs/swagger/swagger.yaml
basePath: /v1
definitions:
  rest.BuildRequest:
    properties:
      version:
        $ref: '#/definitions/v1.SystemVersion'
        type: object
    type: object
...
  rest.DeployRequest:
    properties:
      buildId:
        $ref: '#/definitions/v1.BuildID'
        type: object
      version:
        $ref: '#/definitions/v1.SystemVersion'
        type: object
    type: object
...
  v1.Build:
    properties:
      completionTimestamp:
        type: string
      id:
        $ref: '#/definitions/v1.BuildID'
        type: object
      services:
        description: |
          Services maps service paths (e.g. /foo/bar/buzz) to the
          status of the build for that service in the Build.
        type: object
      startTimestamp:
        type: string
      state:
        $ref: '#/definitions/v1.BuildState'
        type: object
      version:
        $ref: '#/definitions/v1.SystemVersion'
        type: object
    type: object
  v1.BuildID:
    type: object
  v1.BuildState:
    type: object
  v1.Deploy:
    properties:
      buildId:
        $ref: '#/definitions/v1.BuildID'
        type: object
      id:
        $ref: '#/definitions/v1.DeployID'
        type: object
      state:
        $ref: '#/definitions/v1.DeployState'
        type: object
    type: object
  v1.DeployID:
    type: object
  v1.DeployState:
    type: object
  v1.Job:
    properties:
      completionTimestamp:
        type: string
      id:
        $ref: '#/definitions/v1.JobID'
        type: object
      path:
        type: string
      startTimestamp:
        type: string
      state:
        $ref: '#/definitions/v1.JobState'
        type: object
    type: object
  v1.JobID:
    type: object
  v1.JobState:
    type: object
  v1.NodePool:
    properties:
      failure_info:
        $ref: '#/definitions/v1.NodePoolFailureInfo'
        type: object
      id:
        type: string
      instanceType:
        description: |
          FIXME: how to deal with epochs?
        type: string
      numInstances:
        type: integer
      path:
        type: string
      state:
        $ref: '#/definitions/v1.NodePoolState'
        type: object
    type: object
  v1.NodePoolFailureInfo:
    properties:
      message:
        type: string
      time:
        type: string
    type: object
  v1.NodePoolState:
    type: object
...
  v1.SystemVersion:
    type: object
...
host: localhost:8876
info:
  contact: {}
  description: Welcome to lattice api
  license:
    name: Apache 2.0
  termsOfService: TBD
  title: Lattice API
  version: "1.0"
paths:
...
  /systems/{systemId}/builds:
    get:
      consumes:
      - application/json
      description: list builds
      operationId: list-builds
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.Build'
            type: array
      summary: Lists builds
    post:
      consumes:
      - application/json
      description: build system
      operationId: build-system
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      - description: Create build
        in: body
        name: buildRequest
        required: true
        schema:
          $ref: '#/definitions/rest.BuildRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Build'
            type: object
      summary: Build system
  /systems/{systemId}/builds/{id}:
    get:
      consumes:
      - application/json
      description: get build
      operationId: get-build
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      - description: Build ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Build'
            type: object
      summary: Get build
  /systems/{systemId}/builds/{id}/logs:
    get:
      consumes:
      - application/json
      description: get logs
      operationId: get-build-logs
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      - description: Build ID
        in: path
        name: id
        required: true
        type: string
      - description: Node Path
        in: query
        name: path
        required: true
        type: string
      - description: Sidecar
        in: query
        name: sidecar
        type: string
      - description: Follow
        in: query
        name: follow
        type: string
      - description: Previous
        in: query
        name: previous
        type: boolean
      - description: Timestamps
        in: query
        name: timestamps
        type: boolean
      - description: tail
        in: query
        name: tail
        type: integer
      - description: Since
        in: query
        name: since
        type: string
      - description: Since Time
        in: query
        name: sinceTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: log stream
          schema:
            type: string
      summary: Get build logs
  /systems/{systemId}/deploys:
    get:
      consumes:
      - application/json
      description: list deploys
      operationId: list-deploys
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.Deploy'
            type: array
      summary: Lists deploys
    post:
      consumes:
      - application/json
      description: build system
      operationId: deploy-system
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      - description: Create deploy
        in: body
        name: deployRequest
        required: true
        schema:
          $ref: '#/definitions/rest.DeployRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Deploy'
            type: object
      summary: Deploy system
  /systems/{systemId}/deploys/{id}:
    get:
      consumes:
      - application/json
      description: get deploy
      operationId: get-deploy
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      - description: Deploy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Deploy'
            type: object
      summary: Get deploy
  /systems/{systemId}/jobs:
    get:
      consumes:
      - application/json
      description: list jobs
      operationId: list-jobs
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.Job'
            type: array
      summary: Lists jobs
  /systems/{systemId}/jobs/{id}:
    get:
      consumes:
      - application/json
      description: get job
      operationId: get-job
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Job'
            type: object
      summary: Get job
  /systems/{systemId}/jobs/{id}/logs:
    get:
      consumes:
      - application/json
      description: get job logs
      operationId: get-job-logs
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Sidecar
        in: query
        name: sidecar
        type: string
      - description: Follow
        in: query
        name: follow
        type: string
      - description: Previous
        in: query
        name: previous
        type: boolean
      - description: Timestamps
        in: query
        name: timestamps
        type: boolean
      - description: tail
        in: query
        name: tail
        type: integer
      - description: Since
        in: query
        name: since
        type: string
      - description: Since Time
        in: query
        name: sinceTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: log stream
          schema:
            type: string
      summary: Get job logs
  /systems/{systemId}/node-pools:
    get:
      consumes:
      - application/json
      description: list node pools
      operationId: list-node-pools
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.NodePool'
            type: array
      summary: Lists node pools
  /systems/{systemId}/node-pools/{id}:
    get:
      consumes:
      - application/json
      description: get node pool
      operationId: get-node-pool
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      - description: NodePool ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.NodePool'
            type: object
      summary: Get node pool
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: apiKey
    type: apiKey
swagger: "2.0"